indent option on breakable
This commit adds the ability to toggle whether or not a Breakable will
indent when it is printed. In order to achieve this, we delay pushing
anything into the output buffer until flush is called. When it is
called, we iterate through the doc nodes and check at each step if they
fit into the current line (as we did before). Now, however, when a
Breakable is printed, it checks if it should indent itself or not.

These changes open the door for many more changes that can be
accomplished now that printing is delayed until flush is called. Because
of the delay:

* nodes in the doc tree can now change their representation depending on
  whether or not the current group is broken
* we can force all parent groups to break up the tree, instead of just
  the current group. This will allow all of the nodes up the tree to
  break and change their representation as well
* by passing the indent level through the tree as its being printed
  instead of keeping a single variable around, we can more tightly
  control indentation level as the tree is printed.